+++ sdk/anomalydetector/azure-ai-anomalydetector/src/samples/java/com/azure/ai/anomalydetector/ReadmeSamples.java
@@ -14,10 +14,6 @@ import com.azure.core.http.policy.HttpPipelinePolicy;
 import com.azure.identity.DefaultAzureCredentialBuilder;
 
 /**
- * WARNING: MODIFYING THIS FILE WILL REQUIRE CORRESPONDING UPDATES TO README.md FILE. LINE NUMBERS
- * ARE USED TO EXTRACT APPROPRIATE CODE SEGMENTS FROM THIS FILE. ADD NEW CODE AT THE BOTTOM TO AVOID CHANGING
- * LINE NUMBERS OF EXISTING CODE SAMPLES.
- *
  * Code samples for the README.md
  */
 public class ReadmeSamples {
@@ -26,6 +22,7 @@ public class ReadmeSamples {
      * Sample for creating async client.
      */
     public void createAsyncClient() {
+        // BEGIN: readme-sample-createAnomalyDetectorAsyncClient
         String endpoint = "<anomaly-detector-resource-endpoint>";
         HttpHeaders headers = new HttpHeaders()
             .put("Accept", ContentType.APPLICATION_JSON);
@@ -41,12 +38,14 @@ public class ReadmeSamples {
             .pipeline(httpPipeline)
             .endpoint(endpoint)
             .buildAsyncClient();
+        // END: readme-sample-createAnomalyDetectorAsyncClient
     }
 
     /**
      * Sample for creating sync client.
      */
     public void createClient() {
+        // BEGIN: readme-sample-createAnomalyDetectorClient
         String endpoint = "<anomaly-detector-resource-endpoint>";
         HttpHeaders headers = new HttpHeaders()
             .put("Accept", ContentType.APPLICATION_JSON);
@@ -62,5 +61,6 @@ public class ReadmeSamples {
             .pipeline(httpPipeline)
             .endpoint(endpoint)
             .buildClient();
+        // END: readme-sample-createAnomalyDetectorClient
     }
 }