+++ sdk/core/azure-core/src/samples/java/com/azure/core/util/logging/ClientLoggerJavaDocCodeSnippets.java
@@ -73,6 +73,42 @@ public class ClientLoggerJavaDocCodeSnippets {
             logger.error("A formattable message. Hello, {}", name, ex);
         }
         // END: com.azure.core.util.logging.clientlogger.error#string-object
+
+        // BEGIN: com.azure.core.util.logging.clientlogger.atInfo
+        logger.atInfo()
+            .addKeyValue("key", "value")
+            .log("A formattable message. Hello, {}", name);
+        // END: com.azure.core.util.logging.clientlogger.atInfo
+
+        // BEGIN: com.azure.core.util.logging.clientlogger.atWarning
+        logger.atWarning()
+            .addKeyValue("key", "value")
+            .log("A formattable message. Hello, {}", name, exception);
+        // END: com.azure.core.util.logging.clientlogger.atWarning
+
+        // BEGIN: com.azure.core.util.logging.clientlogger.atError#deffered-value
+        try {
+            upload(resource);
+        } catch (IOException ex) {
+            logger.atError()
+                .addKeyValue("key", () -> "Expensive to calculate value")
+                .log("A formattable message. Hello, {}", name, ex);
+        }
+        // END: com.azure.core.util.logging.clientlogger.atError#deffered-value
+
+        // BEGIN: com.azure.core.util.logging.clientlogger.atverbose.addKeyValue#primitive
+        logger.atVerbose()
+            .addKeyValue("key", 1L)
+            .log(() -> String.format("Param 1: %s, Param 2: %s, Param 3: %s", "param1", "param2", "param3"));
+        // END: com.azure.core.util.logging.clientlogger.atverbose.addKeyValue#primitive
+
+        // BEGIN: com.azure.core.util.logging.loggingeventbuilder
+        logger.atInfo()
+            .addKeyValue("key1", "value1")
+            .addKeyValue("key2", true)
+            .addKeyValue("key3", () -> getName())
+            .log("A formattable message. Hello, {}", name);
+        // END: com.azure.core.util.logging.loggingeventbuilder
     }
 
     /**