+++ sdk/resourcemanager/azure-resourcemanager-containerservice/src/main/java/com/azure/resourcemanager/containerservice/implementation/OpenShiftManagedClustersClientImpl.java
@@ -780,7 +780,7 @@ public final class OpenShiftManagedClustersClientImpl
                 this.client.getHttpPipeline(),
                 OpenShiftManagedClusterInner.class,
                 OpenShiftManagedClusterInner.class,
-                Context.NONE);
+                this.client.getContext());
     }
 
     /**
@@ -1053,7 +1053,7 @@ public final class OpenShiftManagedClustersClientImpl
                 this.client.getHttpPipeline(),
                 OpenShiftManagedClusterInner.class,
                 OpenShiftManagedClusterInner.class,
-                Context.NONE);
+                this.client.getContext());
     }
 
     /**
@@ -1303,7 +1303,8 @@ public final class OpenShiftManagedClustersClientImpl
         Mono<Response<Flux<ByteBuffer>>> mono = deleteWithResponseAsync(resourceGroupName, resourceName);
         return this
             .client
-            .<Void, Void>getLroResult(mono, this.client.getHttpPipeline(), Void.class, Void.class, Context.NONE);
+            .<Void, Void>getLroResult(
+                mono, this.client.getHttpPipeline(), Void.class, Void.class, this.client.getContext());
     }
 
     /**