+++ sdk/storage/azure-storage-common/src/main/java/com/azure/storage/common/Utility.java
@@ -30,8 +30,10 @@ public final class Utility {
     private static final ClientLogger LOGGER = new ClientLogger(Utility.class);
     private static final String UTF8_CHARSET = "UTF-8";
 
-    // Please see <a href=https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/azure-services-resource-providers>here</a>
-    // for more information on Azure resource provider namespaces.
+    /**
+     * Please see <a href=https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/azure-services-resource-providers>here</a>
+     * for more information on Azure resource provider namespaces.
+      */
     public static final String STORAGE_TRACING_NAMESPACE_VALUE = "Microsoft.Storage";