+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fixed a bug that would cause authenticating with a sas token to fail if the timestamps in the token were formatted differently.
 
 ### Other Changes
+- Deprecated BlobClient.uploadWithResponse that does not return a response.
 
 ## 12.14.2 (2021-11-10)